requirements.txt
transformers>=4.36.0
torch>=2.0.0
Pillow>=9.0.0
janus-pro>=0.1.0 